+++ src/main.rs
@@ -9,7 +9,7 @@ fn compile_from_cli() -> Result<()> {
     let main_module = compiler.compile()?;
 
     let target = compiler.target()?;
-    if compiler.options.backup {
+    if compiler.config.backup {
         target.write_backup()?;
     }
     target.write(&main_module)?;